Extract TimerServiceRegistry interface from its implementation class.
