+++ ejb3/src/main/java/org/jboss/as/ejb3/timerservice/NonFunctionalTimerService.java
@@ -26,6 +26,7 @@ import java.lang.reflect.Method;
 import java.util.Collection;
 import java.util.Collections;
 import java.util.Date;
+
 import javax.ejb.EJBException;
 import javax.ejb.ScheduleExpression;
 import javax.ejb.Timer;
@@ -34,12 +35,12 @@ import javax.ejb.TimerService;
 
 import org.jboss.as.ee.component.Component;
 import org.jboss.as.ejb3.component.EJBComponentDescription;
-import org.jboss.as.ejb3.component.TimerServiceRegistry;
 import org.jboss.as.ejb3.component.allowedmethods.AllowedMethodsInformation;
 import org.jboss.as.ejb3.component.allowedmethods.MethodType;
 import org.jboss.as.ejb3.component.singleton.SingletonComponent;
 import org.jboss.as.ejb3.context.CurrentInvocationContext;
 import org.jboss.as.ejb3.logging.EjbLogger;
+import org.jboss.as.ejb3.timerservice.spi.TimerServiceRegistry;
 import org.jboss.invocation.InterceptorContext;
 import org.jboss.msc.service.Service;
 import org.jboss.msc.service.ServiceName;
@@ -130,7 +131,7 @@ public final class NonFunctionalTimerService implements TimerService, Service<Ti
 
         // query the registry
         if (this.timerServiceRegistry != null) {
-            return this.timerServiceRegistry.getAllActiveTimers();
+            return this.timerServiceRegistry.getAllTimers();
         }
         // If we don't have the timer service registry (for whatever reason),
         // we just return an empty collection (since this is a non-functional timer service)
